feat+fix: add PricingPreview with Stripe test mode, fullscreen AI chat, and AIChat polish
Payment & Subscription:
- Add PricingPreview component for Stripe Test Mode checkout
- Add PaymentResultModal for success/cancel feedback
- Add payments router with checkout session, subscription status, webhook
- Add user subscription fields to DB model and migration
- Support decimal pricing (e.g. $9.99) in checkout

AI Chat Improvements:
- Add fullscreen immersive chat mode with ESC to exit
- Left panel: expanded conversation area with compact header
- Right panel: stage snapshot, syncable content, next question, quick actions
- Fix duplicate loading states (merged into single placeholder message)
- Fix text overflow in messages (break-words, overflow-x-auto for code blocks)
- Compact header: shrink pet avatar from 80px to 40px, horizontal layout
- Debounced save for MonetizeStage numeric inputs

+++ backend/app/routers/payments.py
@@ -49,7 +49,7 @@ def create_checkout_session(
         price_data: Dict[str, Any] = {
             "currency": "usd",
             "product_data": {"name": item.name},
-            "unit_amount": item.price * 100,  # Stripe 使用美分
+            "unit_amount": int(item.price * 100),  # Stripe 使用美分
         }
 
         if item.period in ("month", "year"):
